[feat] Implement summarizing statstic for each engagement
Since now we are enforcing the customer to fill out a form that scores
different aspects about each iteration, this commit implements a display
of a summarizing statistic that averages all iterations with valid
customer feedback. Valid customer feedback means a customer feedback with
valid JSON format according to the form we created.